# Implement frameless internal function calls