… Zend/tests/frameless_jmp_001.phpt
… Zend/tests/frameless_jmp_002.phpt
… Zend/tests/frameless_jmp_003.phpt
… Zend/tests/frameless_jmp_004.phpt
… Zend/tests/frameless_jmp_005.phpt
… Zend/tests/frameless_throwing_destructor.phpt
… Zend/tests/frameless_undefined_var.phpt
… Zend/zend_API.h
… Zend/zend_builtin_functions_arginfo.h
… Zend/zend_compile.c
… Zend/zend_execute.c
… Zend/zend_execute.h
… Zend/zend_execute_API.c
… Zend/zend_frameless_function.c
… Zend/zend_frameless_function.h
… Zend/zend_opcode.c
… Zend/zend_vm_def.h
… Zend/zend_vm_execute.h
… Zend/zend_vm_gen.php
… Zend/zend_vm_handlers.h
… Zend/zend_vm_opcodes.c
… Zend/zend_vm_opcodes.h
… build/gen_stub.php
… configure.ac
… ext/opcache/jit/zend_jit.c
… ext/opcache/jit/zend_jit_helpers.c
… ext/opcache/jit/zend_jit_ir.c
… ext/opcache/jit/zend_jit_trace.c
… ext/opcache/tests/ct_eval_frameless_001.phpt
… ext/opcache/tests/ct_eval_frameless_002.phpt