- [x] ~~Coercion. Coercion happens in-place. However, since args to frameless functions aren't copied coercing them in-place can escape the function boundary (for VAR|CONST). We should either copy these values in the handler for these op types, or handle the coercion+freeing in the functions.~~ *Edit*: This was only relevant to string for the current frameless handlers. It was solved by creating a temporary zval that may be used if the parameter is not a string. After the function, the zval must be cleaned up if used, which unfortunately requires manual handling in the frameless handler.
- [x] ~~We probably need observer support~~ Frameless calls are disabled when there's an observer extension for now
- JIT support
    - [x] Tracing
    - [x] Function
- [x] Try arity-completion (provide default-args for handlers with larger arity)