@staabm 2023-11-01
Do these frameless functions influence how stack limit works (e.g. max callstack size exceeded)?