@iluuu1994 2023-11-11
@nielsdos I agree. I will change this so that omitting the flags sets it to `-1`. Btw, any idea what's going on with the LLVM script?