@nielsdos 2023-11-11
> Btw, any idea what's going on with the LLVM script?

Huh, it worked yesterday. It complains about the distro version being 22.04.3, but that is the same version as yesterday...
Looking at the script's code, it fails at a wget to `https://apt.llvm.org/jammy/`, but that page loads locally for me...