@iluuu1994 2023-11-11
@nielsdos It gave back a 403 just 10 minutes ago. I suppose they have (or had) server issues.