@derickr 2023-12-07
Could you rebase this against the master branch? So that it picks up the new ZEND_API numbers. I am getting segfaults loading extensions without rebuilding them:

```
derick@gargleblaster:~/dev/php/php-src.git$ gdb --args php -v

(gdb) run
Starting program: /usr/local/php/master/bin/php -v
[Thread debugging using libthread_db enabled]
Using host libthread_db library "/lib/x86_64-linux-gnu/libthread_db.so.1".

Program received signal SIGSEGV, Segmentation fault.
0x0000555555d3b4a7 in zend_register_functions (scope=0x0, functions=0x7ffff7fbabe0 <geospatial_functions>, function_table=0x0, type=1)
    at /home/derick/dev/php/php-src.git/Zend/zend_API.c:2876
2876				if (info->required_num_args == (uintptr_t)-1) {
(gdb) bt
#0  0x0000555555d3b4a7 in zend_register_functions (scope=0x0, functions=0x7ffff7fbabe0 <geospatial_functions>, function_table=0x0, type=1)
    at /home/derick/dev/php/php-src.git/Zend/zend_API.c:2876
#1  0x0000555555d39e8a in zend_register_module_ex (module=0x7ffff7fbb220 <geospatial_module_entry>, module_type=1) at /home/derick/dev/php/php-src.git/Zend/zend_API.c:2537
#2  0x0000555555b3d2c2 in php_load_extension (filename=0x7ffff3a05030 "geospatial", type=1, start_now=0) at /home/derick/dev/php/php-src.git/ext/standard/dl.c:234
#3  0x0000555555c92781 in php_load_php_extension_cb (arg=0x555556e66e70) at /home/derick/dev/php/php-src.git/main/php_ini.c:317
#4  0x0000555555d18829 in zend_llist_apply (l=0x555556e2ea38 <extension_lists+56>, func=0x555555c9275c <php_load_php_extension_cb>) at /home/derick/dev/php/php-src.git/Zend/zend_llist.c:184
#5  0x0000555555c9373e in php_ini_register_extensions () at /home/derick/dev/php/php-src.git/main/php_ini.c:751
#6  0x0000555555c89476 in php_module_startup (sf=0x555556e0cf60 <cli_sapi_module>, additional_module=0x0) at /home/derick/dev/php/php-src.git/main/main.c:2209
#7  0x0000555555eac8dc in php_cli_startup (sapi_module=0x555556e0cf60 <cli_sapi_module>) at /home/derick/dev/php/php-src.git/sapi/cli/php_cli.c:410
#8  0x0000555555eae34c in main (argc=2, argv=0x555556e36c10) at /home/derick/dev/php/php-src.git/sapi/cli/php_cli.c:1307
```

All fine after a recompile, and I can also confirm that there are no problems for Xdebug with this patch!