@dstogov 2023-12-11
Two thoughts related to this PR or about additional optimizations that may be implemented on top of this.

1) Is it possible to extend ``Optimizer/optimize_func_calls.c`` to convert regular calls to frameless? (may be this is not necessary at all)
2) Can we somehow switch to frameless calls when use internal functions in name-space without ``\``? (this is a common case in Symfony). may be only with JIT?

I don't object against the existing implementation.