@iluuu1994 2023-12-11
> 1. Is it possible to extend Optimizer/optimize_func_calls.c to convert regular calls to frameless? (may be this is not necessary at all)

I don't think there is currently a case where we would know at optimization-time that a function call refers to global scope, where we wouldn't also know it at compile-time. In general, this seems possible though.

> 2. Can we somehow switch to frameless calls when use internal functions in name-space without \? (this is a common case in Symfony). may be only with JIT?

I have two ideas that I can pursue.

1. We could implement a new set of INIT/CALL opcodes that use frameless handlers if available, or pushes a stackframe and stores the operands if unavailable. The result can be cached. This may also improve user calls, since it merges the SEND opcodes. We need two opcodes (INIT/CALL) because we need `EX(call)->func` so that we can compile args with `BP_VAR_FUNC_ARG`, and because calls to undefined functions throw before evaluating arguments.
2. We already make some assumptions that a lexical reference to a function does not suddenly change through a new declaration. https://3v4l.org/9Eqr8 The JIT may use check whether a particular function is shadowed at runtime. This is not the same exact assumption, because runtime cache is cleared between requests, whereas the jitted code is not. So essentially the assumption made is that "if the function is shadowed, it must be shadowed in all requests", which might not hold.

I think these two ideas could also be combined, making 2. simpler if we're already in a similar format, just removing the `INIT` opcode and transforming the `CALL` opcode to a frameless call.