@dstogov 2023-12-11
> Both of these ideas are not trivial though, so I suggest investigating them after this PR has been merged.

I agree.

I see that ``ZEND_OP_IS_FRAMELESS_ICALL()`` won't work in cases without explicit ``ZEND_FRAMELESS_ICALL_*`` instructions. (e.g. when JIT would call frameless handler instead of INIT_FCALL+SEND_*_DO_FCALL). So this part of the PR might have to be redesigned later. It would be great to think about this now, but it's also possible to postpone the decision.