@dstogov 2023-12-11
> @dstogov Fair point. Let me try implementing a PoC for idea 1. This will already require changes to `ZEND_OP_IS_FRAMELESS_ICALL` to support the new opcodes, which should be equally applicable to idea 2 (at least if used in combination).

I think idea 1, may be implemented without any redesign, just by bytecode optimization that replaces INIT_FCALL+SEND+SO_FCALL by ``ZEND_FRAMELESS_ICALL_``. May be, I'm wrong.

For the second idea, for frameless function used in name-spaces we might make zend_compile.c to generate two instruction streams at first pace (e.g. first doing  ``ZEND_FRAMELESS_ICALL_`` and the second ``INIT_FCALL+SEND+SO_FCALL``). It seems this should work with the current PR and should provide improvement with and without JIT.