@dstogov 2023-12-11
> I see, this is also a viable option. Then we just need an opcode that chooses the correct path, based on the called function. This might greatly blow-up instructions (because argument compilation is duplicated), and potentially worsen locality, but would also be the easiest to implement. I'll play around with the options mentioned.

These new opcodes should be used only for unqualified function calls, that may lead to frameless internal functions. I don't this this should dramatically increase the code size.