@iluuu1994 2023-12-20
Ok, here are the results.

  | master (s) | direct-calls (s) | diff (%) | direct-calls-unspecialized (s) | diff (%) | unshared-cache (s) | diff (%) | unshared-cache-unspecialized (s) |  diff (%)
-- | -- | -- | -- | -- | -- | -- | -- | -- | --
bench.php | 44.31177 | 43.837141 | -1.0711 | 43.822742 | -1.1036 | 44.574883 | 0.5937 | 43.815165 | -1.1207
bench.php+jit | 14.999982 | 14.990234 | -0.0649 | 15.027456 | 0.1831 | 15.064637 | 0.4310 | 14.993221 | -0.0450
symfony | 30.243476 | 30.05604 | -0.6197 | 30.121743 | -0.4025 | 29.904663 | -1.1202 | 29.908698 | -1.1069
symfony+jit | 29.249913 | 28.96279 | -0.9816 | 29.011227 | -0.8160 | 28.743947 | -1.7298 | 28.729461 | -1.7793
symfony+fqn | 30.252826 | 29.894223 | -1.1853 | 30.029101 | -0.7395 | 29.795143 | -1.5128 | 29.801449 | -1.4920
symfony+fqn+jit | 29.2487 | 28.941631 | -1.0498 | 28.884016 | -1.2468 | 28.906286 | -1.1706 | 29.034026 | -0.7339
  |   |   | -0.8287 |   | -0.6875 |   | -0.7514 |   | -1.0463

From this benchmark, it does seem that the version without the shared cache slot, and without specializations offers the best trade-offs. The improvement isn't huge but it's measurable. Most significantly, it seems the code is faster in all variants, including with non-fqn function calls (i.e. JMP_FRAMELESS). bench.php was performed with more iterations for higher accuracy. As for hot loops I tested this by running [sebastianbergmann/diff](https://github.com/sebastianbergmann/diff) but without https://github.com/sebastianbergmann/diff/pull/118, and by implementing a specialized version of `max(a2)`.

  | master (ms) | unshared-cache-unspecialized (ms) | diff (%)
-- | -- | -- | --
plain | 741.5 | 620.5 | -16.3182
jit | 335.3 | 307.8 | -8.2016
fqn | 670.5 | 574.6 | -14.3027
fqn+jit | 319.3 | 298.5 | -6.5142

As expected, the difference is bigger when frameless calls may be used in hot loops.

@dstogov Do you think this is enough to justify the added complexity?